requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0
watchdog>=2.1.6
tiktoken>=0.5.0 